fix: harden IMU read path, surface failures, simplify interfaces
Addresses the code-review findings on the roadmap branch:

- IMU interface now reports transport errors: readStepCount() returns
  bool and only the minimal begin()/readStepCount() surface remains
  (writeRegister/readRegister were board-only details, not the contract).
- Pedometer preserves the running total across a failed read instead of
  overwriting it with 0; adds readOk() so callers can detect I2C errors.
- HardwareIMU: writeRegister() returns the endTransmission() status; begin()
  folds in the pedometer config and fails closed if setup is rejected;
  readStepCount() does a single repeated-start burst read of both count
  bytes and returns false on a partial read.
- main.cpp: drop the blocking while(1) halt (return from setup and let
  loop() surface a throttled diagnostic when not LOGGING); add a per-poll
  read-error warning; remove the unused elapsed_timer include.
- StateMachine: remove the unreachable IDLE state, consolidate BOOT->LOGGING
  into startLogging(), extract shared enterLogging().
- Tests: relocate to test/test_nanowear/ (a test_*-prefixed dir) so PIO
  discovers the suite explicitly and adding more suites is additive; add a
  read-failure regression test.
- Cleanups: drop dead setInterval(); simplify redundant step_codec cast;
  update CLAUDE.md/platformio.ini to the real test layout.

+++ src/step_codec.h
@@ -9,7 +9,9 @@
 // Isolated as a pure function so the byte-order maths is unit-testable without
 // touching I2C. Mirrors the original expression `(highByte << 8) | lowByte`.
 inline uint16_t combineStepBytes(uint8_t low, uint8_t high) {
-    return static_cast<uint16_t>((static_cast<uint16_t>(high) << 8) | low);
+    // `high` is uint8_t and is promoted to int before the shift, so only the
+    // final cast to uint16_t is load-bearing.
+    return static_cast<uint16_t>((high << 8) | low);
 }
 
 #endif // NANOWEAR_STEP_CODEC_H
